% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateData}
\alias{generateData}
\title{Generate the original data for simulation}
\usage{
generateData(n)
}
\arguments{
\item{n}{the sample size in each simulated dataset}
}
\value{
the simulated dataset according to the settings in Kontopantelis and Evangelos (2017)
}
\description{
Generate the original data for simulation
}
