% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAR_MethodB.R
\name{MAR_MethodB}
\alias{MAR_MethodB}
\title{MAR_MethodB}
\usage{
MAR_MethodB(n, NSIM, missRate, trueValue, cores = 1)
}
\arguments{
\item{n}{sample size in each simulated dataset}

\item{NSIM}{the number of simulation runs}

\item{missRate}{the missing rate}

\item{trueValue}{the true value of the parameter}

\item{cores}{the number of cores for parallelization, defalut = 1}
}
\value{
the beta, betaSE, meanBias, meanError, cover for methodB
}
\description{
MAR_MethodB
}
