% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAR_MethodD.R
\name{MAR_MethodD}
\alias{MAR_MethodD}
\title{MAR_MethodD}
\usage{
MAR_MethodD(n, NSIM, missRate, trueValue, cores = 1)
}
\arguments{
\item{n}{sample size in each simulated dataset}

\item{NSIM}{the number of simulation runs}

\item{missRate}{the missing rate}

\item{trueValue}{the true value of the parameter}

\item{cores}{the number of cores for parallelization, defalut = 1}
}
\value{
the beta, betaSE, meanBias, meanError, cover for methodD
}
\description{
MAR_MethodD
}
