% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCAR.R
\name{MCAR}
\alias{MCAR}
\title{MCAR}
\usage{
MCAR(missRate, full)
}
\arguments{
\item{missRate}{the missing rate for each variable}

\item{full}{the orignial true dataset}
}
\value{
the simulated MCAR dataset
}
\description{
Generate missing value in the dataset under MCAR
}
