% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCAR_0.8.R
\docType{data}
\name{MCAR_0.8}
\alias{MCAR_0.8}
\title{MCAR_0.8}
\format{The following four methods are evaluated based on the mean bias, mean error, relative bias, coverage, mean square error, standard deviation.
\describe{
\item{Method A}{Complete case analysis}
\item{Method B}{No outcome imputation, not included in mi model}
\item{Method C}{No outcome imputation, outcome included in mi model}
\item{Method D}{Outcome imputed and included in mi model}
}}
\usage{
MCAR_0.8
}
\description{
Summarized simulation results for coefficient of treatment using method A-D under MCAR with 0.8 missing rate.
}
\keyword{datasets}
