% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCAR_MethodA.R
\name{MCAR_MethodA}
\alias{MCAR_MethodA}
\title{MCAR_MethodA}
\usage{
MCAR_MethodA(n, NSIM, missRate, trueValue, cores = 1)
}
\arguments{
\item{n}{sample size in each simulated dataset}

\item{NSIM}{the number of simulation runs}

\item{missRate}{the missing rate}

\item{trueValue}{the true value of the parameter}

\item{cores}{the number of cores for parallelization, defalut = 1}
}
\value{
the beta, betaSE, meanBias, meanError, cover for methodA
}
\description{
MCAR_MethodA
}
