% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodA.R
\name{methodA}
\alias{methodA}
\title{methodA: complete case analysis}
\usage{
methodA(dataset)
}
\arguments{
\item{dataset}{the dataset for analysis}
}
\value{
the summary of the model fit
}
\description{
methodA: complete case analysis
}
