% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodB.R
\name{methodB}
\alias{methodB}
\title{method B}
\usage{
methodB(dataset)
}
\arguments{
\item{dataset}{the dataset for analysis}
}
\value{
the summary of the model fit
}
\description{
Method B used in Kontopantelis and Evangelos (2017) for anaylsis incomplete data: no outcome imputation, not included in mi model
}
