% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodC.R
\name{methodC}
\alias{methodC}
\title{method C: no outcome imputation, outcome imputed in mi model}
\usage{
methodC(dataset)
}
\arguments{
\item{dataset}{the dataset for analysis}
}
\value{
the summary of the model fit
}
\description{
method C: no outcome imputation, outcome imputed in mi model
}
