% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_MAR.R
\name{sim_MAR}
\alias{sim_MAR}
\title{sim_MAR}
\usage{
sim_MAR(n, NSIM, missRate, trueValue, cores = 1)
}
\arguments{
\item{n}{sample size in each simulated dataset}

\item{NSIM}{the number of simulation runs}

\item{missRate}{the missing rate}

\item{trueValue}{the true value of the parameter}

\item{cores}{the number of cores for parallelization, defalut = 1}
}
\value{
The summary simulation results
}
\description{
sim_MAR
}
\examples{
\dontrun{
sim_MAR <- function(n = 1000, NSIM = 1000, missRate = 0.2, trueValue = log(2), cores = 1)
}
}
