% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_MCAR.R
\name{sim_MCAR}
\alias{sim_MCAR}
\title{sim_MCAR}
\usage{
sim_MCAR(n, NSIM, missRate, trueValue, cores = 1)
}
\arguments{
\item{n}{sample size in each simulated dataset}

\item{NSIM}{the number of simulation runs}

\item{missRate}{the missing rate}

\item{trueValue}{the true value of the parameter}

\item{cores}{the number of cores for parallelization, defalut = 1}
}
\value{
The summary simulation results
}
\description{
sim_MCAR
}
\examples{
\dontrun{
sim_MCAR(n = 1000, NSIM = 1000, missRate = 0.2, trueValue = log(2), cores = 1)
}
}
