% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAR.R
\name{MAR}
\alias{MAR}
\title{Generate missing values in the dataset under MAR}
\usage{
MAR(missRate, full)
}
\arguments{
\item{missRate}{the missing rate for each variable}

\item{full}{the orignial dataset}
}
\value{
the simulated MCAR dataset
}
\description{
Generate missing values in the dataset under MAR
}
