% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCAR_MethodB.R
\name{MCAR_MethodB}
\alias{MCAR_MethodB}
\title{MCAR_MethodB}
\usage{
MCAR_MethodB(n, NSIM, missRate, trueValue, cores = 1)
}
\arguments{
\item{n}{sample size in each simulated dataset}

\item{NSIM}{the number of simulation runs}

\item{missRate}{the missing rate}

\item{trueValue}{the true value of the parameter}

\item{cores}{the number of cores for parallelization, defalut = 1}
}
\value{
the beta, betaSE, meanBias, meanError, cover for methodB
}
\description{
This function will call the \code{generateData} and \code{MCAR} function to generate the dataset under MCAR with certain missing rate,use method B for analysis, repeate for 1000 iterations and return the beta, betaSE, meanBias, meanError, cover for each iteratio
}
