% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodB.R
\name{methodB}
\alias{methodB}
\title{method B: no outcome imputation, not included in mi model}
\usage{
methodB(dataset)
}
\arguments{
\item{dataset}{the dataset for analysis}
}
\value{
the summary of the model fit
}
\description{
method B: no outcome imputation, not included in mi model
}
