% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodC.R
\name{methodC}
\alias{methodC}
\title{method C}
\usage{
methodC(dataset)
}
\arguments{
\item{dataset}{the dataset for analysis}
}
\value{
the summary of the model fit
}
\description{
Method C used in Kontopantelis and Evangelos (2017) for anaylsis incomplete data: no outcome imputation, outcome included in mi model
}
