% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodD.R
\name{methodD}
\alias{methodD}
\title{method D: outcome imputed and included in mi model}
\usage{
methodD(dataset)
}
\arguments{
\item{dataset}{the dataset for analysis}
}
\value{
the summary of the model fit
}
\description{
method D: outcome imputed and included in mi model
}
