% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_MAR.R
\name{sim_MAR}
\alias{sim_MAR}
\title{sim_MAR}
\usage{
sim_MAR(n, NSIM, missRate, trueValue, cores = 1)
}
\arguments{
\item{n}{sample size in each simulated dataset}

\item{NSIM}{the number of simulation runs}

\item{missRate}{the missing rate}

\item{trueValue}{the true value of the parameter}

\item{cores}{the number of cores for parallelization, defalut = 1}
}
\value{
The summary simulation results
}
\description{
This function will call the function \code{MAR_MethodA}, \code{MAR_MethodB}, \code{MAR_MethodC} and \code{MAR_MethodD}
to generate the beta coeffients for each iteration and return the summarized simulation results including mean bias, mean error,
relative bias, coverage, mean square error, standard deviation
}
\examples{
\dontrun{
sim_MAR(n = 1000, NSIM = 1000, missRate = 0.2, trueValue = log(2), cores = 1)
}
}
