% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_MCAR.R
\name{sim_MCAR}
\alias{sim_MCAR}
\title{Comparing the Performance of Method A-D (MCAR)}
\usage{
sim_MCAR(n, NSIM, missRate, trueValue, cores = 1)
}
\arguments{
\item{n}{sample size in each simulated dataset}

\item{NSIM}{the number of simulation runs}

\item{missRate}{the missing rate}

\item{trueValue}{the true value of the parameter}

\item{cores}{the number of cores for parallelization, defalut = 1}
}
\value{
The summary simulation results
}
\description{
This function will call the function \code{MCAR_MethodA}, \code{MCAR_MethodB}, \code{MCAR_MethodC} and \code{MCAR_MethodD}
to generate the beta coeffients for each iteration and return the summarized simulation results including mean bias, mean error,
relative bias, coverage, mean square error, standard deviation to compare the methods
}
\examples{
\dontrun{
sim_MCAR(n = 1000, NSIM = 1000, missRate = 0.2, trueValue = log(2), cores = 1)
}
}
